lattice_decoder common functionality, rep code decoder
