+++ topological_codes/__init__.py
@@ -7,4 +7,5 @@ from .circuits.xzzx import XZZXQubit
 from .circuits.circ import TopologicalRegister, TopologicalCircuit
 
 from .fitters.rotated_surface import RotatedGraphDecoder
+from .fitters.repetition import RepetitionGraphDecoder
